refactor(box-template): delete the provision receiving half — nothing writes the other end
Five files implemented the receiving end of a provisioning concept with no
producer anywhere: a disk reader (provision.mjs), an HTML placeholder, a
server splice, a dev-server plugin and a page parser.

Nothing in this repo — or in the console (vendo-web, checked: it writes only
/exec/bundle.mjs and /exec/request.json into a box) — writes
/app/.vendo/host/theme.json or sets VENDO_APP_ID. The contract lists
VENDO_APP_ID as a cut field. So readTheme() always returned undefined and
applyProvisionedBrand() always returned {}. readHostComponents had no importer
at all.

The theme route that actually works is `?vendoTheme=`, and it is untouched:
src/provision.ts becomes src/theme.ts carrying only the query reader, with the
same isObject/JSON.parse tolerance and the same themeCssVariables mapping.
The appId override goes with it — its stated justification named the
provider-URL branch, which was deleted (served-orgs.test.ts:93 records the
handoff the org branch now shares "the same ... the personal branch gets").

No changeset: @vendoai/box-template is private and privatePackages.version is
false. No tests deleted — nothing covered the cut code.
