+++ packages/box-template/index.html
@@ -4,11 +4,6 @@
 <meta charset="utf-8">
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <title>Vendo app</title>
-<!-- Provision data, spliced in by whichever server serves this page (server.js
-     in production, the dev server's plugin for the live preview). Read at
-     RUNTIME on purpose: which app this is and whose brand it wears are per-box
-     facts, and a Vite build would freeze them into the template. -->
-<script id="vendo-runtime" type="application/json">__VENDO_RUNTIME__</script>
 <style>
   *, *::before, *::after { box-sizing: border-box; }
   html, body { margin: 0; }
