+++ packages/box-template/tsconfig.json
@@ -16,5 +16,5 @@
     "checkJs": false,
     "types": ["node", "vite/client"]
   },
-  "include": ["src", "vite.config.ts", "provision.mjs", "scripts"]
+  "include": ["src", "vite.config.ts", "scripts"]
 }
